requirements.txt
asgiref==3.2.3
Django==3.0.3
django-rest-authtoken==1.2.4
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
